Make sure the gameController is getting the player instances from menu
I don't need so many parameters when the menu is a single instance whose
state is available. Therefore players do not need to be passed around
as parameters constantly as that makes it harder to maintain. Getting
players and passing them around shouldn't be that hard and we shouldn't
need to pass that much anyway to reduce dependencies. We reduce
dependencies by adding the global variable and I'm not sure if that is
much better. It does make the code easier I think.