Reorganize code for multiple instance vs single instance
Factory functions for multiple instances are at the top and single
instance modules are at the bottom, i.e. reuse is not necessary.
We don't need to create multiple objects in memory for the application
for it. This makes it easier to follow to eventually get to the main
entity responsible for controlling the application. This idea is
noted further in my notes and personal documentation.