Refactor factories as modules with exposed set/reset functionality
Instead of creating new instances of controllers across games, instead
use one instance as I wanted that does not set data implicitly,
internally, but rather explicitly through exposure of its method. The
object is thrown away, it's data is just reset through a function on use
for the new round. The needed unreset data is taken care of and cascades
through the functionality that works with the data, and should never
be set by default on the first initialization but rather explicitly set.

Furthermore I combined modules with inheritance since that is possible.
See stored documentation notes for expansion and why my assumptions were
incorrect.