Allow for multiple human-bot rounds play through DOM
This refactor was motivated by the following:

The play of multiple rounds is promoted by the loop of human-bot turns
which in this case the loop is simulated by the event listener click
on the UI board obtainable from the services provided by the display
controller which tracks all these aspects of these components. The play
of multiple rounds continues until the termination signal of the board
being filled, or alternatively either the player or bot wins. The latter
is yet to be implemented, but for now the game should definitely
terminate.