Refactor play round into removable play round (which fixes nothing)
This fixes nothing and causes the problem I was trying to avoid. I
cannot pass the event object from the event listener to the called
function. Function script can be invoked receiving the event object
though I am sure, that will have to be my theory for now and we can test
it be testing whether the stored row and column are accurate.