Move remove event listener to where play round reference is accessible
There are 3 controllers with different play rounds and I cannot.
I wish I could enclose this logic within the unified function but it
does not seem possible. I thought you just need the name of the function
not the function script itself. I could of course pass the reference to
the play round function to remove the event listener on. This is
slightly messy as it uses parameters but is still arguably cleaner and
more maintainable since all repeatable common code is in one place.