Fix bug after bot wins in human-bot game, human can play again
In debugging I discovered that the game was not read as complete when
the next bot move displayed was a winning move, and the board by then
should have reflected it was the winning move. This was all because the
condition gets evaluated once and the result is stored in the variable.
The expression is not evaluated multiple times. It is evaluated and its
result assigned to the boolean variable.