Try add get and set winner functions
This prevents memorizing when the active player is the winner. It is
known by the get winner function. Either get the last active player or
get winner as set in the gameplayController so we set it just right. I
realize either one works since in the gameResults we know that that last
active player is the winner, of course we can set the winner even
earlier than in this end game function but its fine. We will use the
knowledge we have in the end game function.